+++ crates/near-contract-transport/src/views/observation.rs
